+++ packages/core-mobile/app/new/ContextApp.tsx
@@ -1,12 +1,10 @@
 /**
  * Context wrapper for App
  **/
-import React, { FC, PropsWithChildren, useEffect, useState } from 'react'
+import React, { FC, PropsWithChildren } from 'react'
 import * as Sentry from '@sentry/react-native'
 import Toast from 'react-native-toast-notifications'
-import JailMonkey from 'jail-monkey'
 import { RootSiblingParent } from 'react-native-root-siblings'
-import JailbrokenWarning from 'screens/onboarding/JailbrokenWarning'
 import { PosthogContextProvider } from 'contexts/PosthogContext'
 import { EncryptedStoreProvider } from 'contexts/EncryptedStoreProvider'
 import { ReactQueryProvider } from 'contexts/ReactQueryProvider'
@@ -14,6 +12,7 @@ import { DeeplinkContextProvider } from 'contexts/DeeplinkContext/DeeplinkContex
 import SentryService from 'services/sentry/SentryService'
 import TopLevelErrorFallback from './common/components/TopLevelErrorFallback'
 import { App } from './App'
+import JailbreakCheck from './common/components/JailbreakCheck'
 
 function setToast(toast: Toast): void {
   global.toast = toast
@@ -37,11 +36,10 @@ const ContextApp = (): JSX.Element => {
   return (
     <Sentry.ErrorBoundary fallback={<TopLevelErrorFallback />}>
       <ContextProviders>
-        <JailBrokenCheck>
-          <RootSiblingParent>
-            <App />
-          </RootSiblingParent>
-        </JailBrokenCheck>
+        <RootSiblingParent>
+          <App />
+        </RootSiblingParent>
+        <JailbreakCheck />
         <Toast
           ref={ref => {
             ref && setToast(ref)
@@ -52,23 +50,6 @@ const ContextApp = (): JSX.Element => {
       </ContextProviders>
     </Sentry.ErrorBoundary>
   )
-}
-
-const JailBrokenCheck: FC<PropsWithChildren> = ({ children }) => {
-  const [showJailBroken, setShowJailBroken] = useState(false)
-
-  useEffect(() => {
-    if (!__DEV__ && JailMonkey.isJailBroken()) {
-      setShowJailBroken(true)
-    }
-  }, [])
-
-  if (showJailBroken) {
-    // TODO: convert JailbrokenWarning to new design
-    return <JailbrokenWarning onOK={() => setShowJailBroken(false)} />
-  }
-
-  return <>{children}</>
 }
 
 export default SentryService.isAvailable ? Sentry.wrap(ContextApp) : ContextApp
