+++ packages/core-mobile/app/new/common/components/TopLevelErrorFallback.tsx
@@ -1,18 +1,7 @@
-import { K2AlpineThemeProvider } from '@avalabs/k2-alpine'
 import React from 'react'
-import { useColorScheme } from 'react-native'
 import RNRestart from 'react-native-restart'
+import { withK2AlpineThemeProvider } from './withK2AlpineThemeProvider'
 import { GenericWarning } from './GenericWarning'
-
-const withK2AlpineThemeProvider = (Component: React.ComponentType) => () => {
-  const colorScheme = useColorScheme()
-
-  return (
-    <K2AlpineThemeProvider colorScheme={colorScheme}>
-      <Component />
-    </K2AlpineThemeProvider>
-  )
-}
 
 const TopLevelErrorFallback = (): JSX.Element => {
   return (
